logging
Extracted logging module
